Move LlmExtractor to an exploratory-tier module
extractor.py declared EPISTEMIC_TIER = "pipelines" but held
LlmExtractor, which performs model inference. The pipelines contract
in docs/steering/epistemic-tiers.md forbids inference, and the
package's pipelines-tier __init__ re-exported the class.

The deterministic parts stay in extractor.py: the ExtractionInput and
ExtractionVerdict types, the SnippetExtractor protocol,
KeywordExtractor, and the JSON payload parsing and pair validation.
The prompt, ChatClient and ChatFn, build_user_prompt, LlmExtractor,
build_llm_extractor, and the OpenAI wiring move to llm_extractor.py,
declared exploratory. The package no longer re-exports LlmExtractor or
build_llm_extractor; extractor_eval.py (already exploratory), the
tests, the notebook, and the write-up import from the new module.

check_tier_boundaries.py now enforces the boundary: a future
pipelines-tier import of the LLM path fails CI, which is the intended
consequence of the split.
